TraceViewController now controls BranchNavigator and ExpressionDataExplorer view-models
